+++ src/run_weidu.rs
@@ -10,7 +10,7 @@ use regex::Regex;
 use crate::args::Install;
 use crate::global::Global;
 use crate::module::language::{LanguageOption, LanguageSelection, select_language};
-use crate::components::{Component, Components};
+use crate::module::components::{Component, Components};
 use crate::lowercase::LwcString;
 use crate::module::weidu_mod::{WeiduMod, BareMod};
 use crate::run_result::RunResult;
