+++ src/sub/extract_manifest.rs
@@ -7,7 +7,7 @@ use anyhow::{Result, bail};
 
 use crate::args::Reverse;
 use crate::canon_path::CanonPath;
-use crate::components::FullComponent;
+use crate::module::components::FullComponent;
 use crate::global::Global;
 use crate::log_parser::{parse_weidu_log, LogRow};
 use crate::lowercase::lwc;
