+++ src/sub/reset.rs
@@ -5,7 +5,7 @@ use log::info;
 
 use crate::args::Reset;
 use crate::canon_path::CanonPath;
-use crate::components::Components;
+use crate::module::components::Components;
 use crate::module::manifest::Manifest;
 use crate::module::module::Module;
 use crate::run_weidu::run_weidu_uninstall;
